Deduplicate copyFileTo
