+++ packages/admin-ui/src/__mock-backend__/index.ts
@@ -63,7 +63,7 @@ export const otherCalls: Omit<Routes, keyof typeof namedSpacedCalls> = {
   changeIpfsTimeout: async () => {},
   cleanCache: async () => {},
   cleanDb: async () => {},
-  copyFileTo: async () => {},
+  copyFileToDockerContainer: async () => {},
   diagnose: async () => [],
   ethClientFallbackSet: async () => {},
   ethClientTargetSet: async () => {},
