+++ packages/dappmanager/package.json
@@ -64,7 +64,6 @@
     "compression": "^1.7.4",
     "cookie-session": "^1.4.0",
     "cors": "^2.8.5",
-    "data-uri-to-buffer": "^2.0.0",
     "deepmerge": "^2.2.1",
     "ethers": "5.7.2",
     "express": "^4.17.1",
