+++ packages/dockerApi/package.json
@@ -25,6 +25,7 @@
     "@dappnode/utils": "^0.1.0",
     "@types/lodash-es": "^4.17.9",
     "dargs": "7.0.0",
+    "data-uri-to-buffer": "^2.0.0",
     "dockerode": "^3.2.1",
     "lodash-es": "^4.17.21",
     "memoizee": "0.4.14",
