+++ packages/dockerApi/src/getDockerImageManifest.ts
@@ -5,22 +5,8 @@ interface ImageManifest {
   Layers: string[]; // ["14ec119e6215a169a53a8c9cdfb56ca873e10f2e5ea0a37692bfa71601f18ec7/layer.tar"]
   RepoTags: string[]; // ["package.dnp.dappnode.eth:0.2.0"];
 }
-export function dockerImageManifest(
+export async function getDockerImageManifest(
   imagePath: string
 ): Promise<ImageManifest[]> {
   return shell(`tar -xOf ${imagePath} manifest.json`).then(JSON.parse);
-}
-
-// File manager, copy command
-
-export function dockerCopyFileTo(
-  id: string,
-  fromPath: string,
-  toPath: string
-): Promise<string> {
-  return shell(`docker cp --follow-link ${fromPath} ${id}:${toPath}`);
-}
-
-export function dockerGetContainerWorkingDir(id: string): Promise<string> {
-  return shell(`docker inspect --format='{{json .Config.WorkingDir}}' ${id}`);
 }
