+++ packages/installer/package.json
@@ -31,7 +31,6 @@
     "@dappnode/schemas": "^0.1.0",
     "@dappnode/toolkit": "^0.1.0",
     "@dappnode/utils": "^0.1.0",
-    "data-uri-to-buffer": "2.0.0",
     "deepmerge": "2.2.1",
     "dockerode": "3.2.1",
     "ethers": "^6.9.0",
